Added Frontmatter in Preview (Dataview + CSS)
Made a really complicated and long Dataview Query that puts a file's frontmatter into a table in a janky way and then adjusts the HTML manually.

Also added some shoddy CSS to the whole thing.

Did this as a proof-of-concept, since i am not really all that proficient in DQL and wanted to see if it's possible first.