Improved Frontmatter in Preview (Dataview + CSS)
Drastically reworked the whole thing, after researching how to properly display the frontmatter in a table without even querying.

Adjusted the CSS accordingly. This came at the cost of styling the tags, as the current dataviewjs code block is just two lines long and i din't want to add bulky js just to style the tags a little.